I init API endpoint to getCollaborators
