+++ .env.example
@@ -1,2 +1,3 @@
 NEXT_PUBLIC_ENABLE_TESTNETS=
 NEXT_PUBLIC_NFT_STORAGE_TOKEN=
+GITHUB_PERSONAL_ACCESS_TOKEN=
