+++ README.md
@@ -16,6 +16,10 @@ Then, run the development server:
 npm run dev
 ```
 
+make sure you set up the ENV variables:
+
+- GITHUB_PERSONAL_ACCESS_TOKEN - used to get Active Developers from Github.
+
 Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
 
 You can start editing the page by modifying `pages/index.tsx`. The page auto-updates as you edit the file.
