+++ components/Developers/index.tsx
@@ -1,0 +1,3 @@
+import Developers from "./Developers";
+
+export default Developers;
