+++ package.json
@@ -13,6 +13,7 @@
     "@hookform/error-message": "^2.0.0",
     "@hookform/resolvers": "^2.9.8",
     "@rainbow-me/rainbowkit": "^0.7.0",
+    "axios": "^1.2.5",
     "ethers": "^5.0.0",
     "next": "^13.0.0",
     "react": "^18.2.0",
